feat: Translating some ui texts
