+++ src/pages/Home/Home.jsx
@@ -1,0 +1,9 @@
+const Home = () => {
+  return (
+    <>
+      <main></main>
+    </>
+  );
+};
+
+export default Home;
